pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autogensocial-contracts"
version = "0.1.0"
description = "API contracts and data schemas for the AutoGenSocial platform"
readme = "README.md"
license = "MIT"
authors = [
    {name = "1084 Ventures", email = "info@1084ventures.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0.0",
    "pydantic[email]>=2.0.0",
    "typing-extensions>=4.0.0",
    "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "datamodel-code-generator>=0.25.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/1084-Ventures/autogensocial-contracts"
Repository = "https://github.com/1084-Ventures/autogensocial-contracts"
Issues = "https://github.com/1084-Ventures/autogensocial-contracts/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["autogensocial_contracts*"]
exclude = ["tests*", "scripts*"]

[tool.setuptools.package-data]
"autogensocial_contracts.openapi" = ["*.yaml", "*.yml"]
"autogensocial_contracts.schemas" = ["*.json"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["autogensocial_contracts"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=autogensocial_contracts",
    "--cov-report=term-missing",
    "--cov-report=html"
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
]

[tool.coverage.run]
source = ["autogensocial_contracts"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*"
]